Count samples with 64 bits in more places
Change a number of places where size_t or off_t were still used to
count samples in order to now use (sox_)uint64_t. Also use PRIu64
where such numbers are output.

Some function prototypes and data structures from sox.h are also
affected. This breaks ABI compatibility; the library version number
has been incremented for 14.4.0 already in commit e9a66b7c, so no
change necessary here.

+++ src/crop.c
@@ -32,7 +32,7 @@ static int parse(sox_effect_t * effp, char * * argv, sox_rate_t rate)
 {
   priv_t * p = (priv_t *)effp->priv;
   char const * s, * q;
-  size_t samples;
+  uint64_t samples;
   int i;
 
   for (i = p->argc - 1; i == 0 || i == 1; --i) {
@@ -105,7 +105,7 @@ static int flow(sox_effect_t * effp, sox_sample_t const * ibuf,
     sox_sample_t * obuf, size_t * isamp, size_t * osamp)
 {
   priv_t * p = (priv_t *)effp->priv;
-  size_t skipped;
+  uint64_t skipped;
 
   p->pos[0].at -= skipped = min(p->pos[0].at, *isamp);
   *osamp = !p->pos[0].at * min(p->pos[1].at, min(*isamp - skipped, *osamp));
